MapKit Demo (#8)
* Add City struct and update UI functionality

The commit introduces a City struct with properties of a city and methods for serialization and deserialization. Also, it removes the placeholder function for city display and replaces it with actual GeoJSON data. Lastly, it updates the UI workflow for creating and displaying cities.

* Refactor City struct, add tests and improve serialization process.

The City struct has now been refactored to make it Codable, replacing the manual serialization and deserialization methods with the functionalities provided by Swift's JSONEncoder/Decoder. Coordinate struct has also been added to replace CLLocationCoordinate2D for better compatibility with JSON encoding/decoding. Unit tests have been added to ensure that serialization and deserialization work accurately. These changes aim for cleaner code, improved performance, and ensure strong correctness of the serialization and deserialization process.