Update app with authentication, map improvements and workflow addition.
Implemented the AuthenticationServices and added an AuthenticationManager for sign in and sign up functionality. The map view was improved with additional features and gestures. A GitHub workflow was also enabled to build and analyze the default scheme using xcodebuild command. All other minor changes and adjustments made were to support these main enhancements.