Webauthn - UI (#9)
* Update app with authentication, map improvements and workflow addition.

Implemented the AuthenticationServices and added an AuthenticationManager for sign in and sign up functionality. The map view was improved with additional features and gestures. A GitHub workflow was also enabled to build and analyze the default scheme using xcodebuild command. All other minor changes and adjustments made were to support these main enhancements.

* Update swift.yaml to add directory listing

In the swift.yaml GitHub workflow file, a command has been added to list the files in the Arkavo directory after changing to it. This gives us more granular visibility into the directory structure during the workflow execution.

* Simplify GitHub workflow in swift.yaml

The swift.yaml GitHub workflow has been simplified by removing the step of explicitly changing to the Xcode project directory. This change condenses the action to make it more efficient. The directory change has been integrated into other steps to ensure the workflow still functions correctly.

* Update swift.yaml to remove Package.resolved file

The swift.yaml workflow file has been updated to include an action that removes the Package.resolved file. This is done for cleaning the project workspace prior to setting the default scheme, ensuring that there are no conflicts with previous resolutions.

* Update GitHub workflows for Swift

This commit adds a step to the workflow file for setting up the Xcode environment. It uses the maxim-lobanov/setup-xcode action with the latest stable version of Xcode. This change ensures that the workflow always uses the most recent stable Xcode version.

* Remove setting of default scheme in Swift workflow

The code for setting the default scheme has been commented out in the Swift GitHub actions workflow. This change does not delete the previous solution, but deactivates it for now. The specific reasons and potential reactivation of this function should be considered based on the actual project needs.

* Update scheme in Swift CI workflow

Changed the scheme name from 'default' to 'Arkavo' in the Swift CI configuration file (.github/workflows/swift.yaml). This alters the environment for the 'Build' action, ensuring that it now targets the 'Arkavo' scheme instead of the default one.

* Remove analyze command from xcodebuild in swift.yaml

The 'analyze' command was removed from the xcodebuild command in the .github/workflows/swift.yaml file, simplifying the build process. This change helps avoid potential issues that might have been caused by the analysis step during the project's build.

* Update swift.yaml to switch from build to test

The swift.yaml workflow configuration file has been updated. The changes remove the build step and instead implement a test step. The update also simplifies the process by directly calling xcodebuild with test and removing the need to set a default scheme.

* Remove swift.yaml GitHub action

The swift.yaml GitHub workflows file, which was responsible for setting up the Xcode build and test environment for iOS, has been deleted. This change signifies a restructuring of continuous integration strategy, as the corresponding build and test setup steps are no longer required in the current project context.

* Remove outdated comments from signIn function

The unused comments in the signIn() function of the AuthenticationManager.swift file were removed. These comments were providing outdated instructions which are no longer applicable or necessary. This clean-up improves code readability and removes confusion.