Implement cluster annotations on map view
The changes introduce a new way of handling map markers. Instead of displaying individual annotations for each city, the cities are now clustered by continent. This is done in order to optimize the map view for a large number of elements by reducing the number of visible markers. Clusters appear as a single marker on the map, which represents multiple cities in the same location. The code for adding and removing cities to and from clusters, updating annotations, and calculating center coordinates for clusters has also been added.