Map Overlay and Clustering (#11)
* Add CityInfoOverlay and refactor city information display

Created a new `CityInfoOverlay` SwiftUI view to encapsulate the display of city information. Refactored the ContentView by moving the annotation and city information logic to their own designated classes, `AnnotationManager` and `NanoTDFManager` respectively. This change improves code readability and separation of concerns.

* Implement cluster annotations on map view

The changes introduce a new way of handling map markers. Instead of displaying individual annotations for each city, the cities are now clustered by continent. This is done in order to optimize the map view for a large number of elements by reducing the number of visible markers. Clusters appear as a single marker on the map, which represents multiple cities in the same location. The code for adding and removing cities to and from clusters, updating annotations, and calculating center coordinates for clusters has also been added.