Update in-app text and rearrange plist strings
The code changes include updates in the application's text, specifically within the 'CityInfoOverlay.swift' and 'CityLabel.swift' files. The 'Info.plist' file was also rearranged, and a usage description for Face ID authentication has been added in the 'project.pbxproj' file. Lastly, a minor adjustment was made in the 'Arkavo.entitlements' file to improve readability.