Save Profile to Database
Now when you Create a Profile it is saved to the database. The profile is added to the account as before so that you can only View Profile after that. I added a @Query so that when you start up, it checks if there is a profile and updates the local account and thought stream.