Stream persist (#38)
* Enhance Stream and Account models with SwiftData integration

Refactor Stream and Account models to use SwiftData and streamline their initializations. Simplify profile creation and handling in `PersistenceController` and update `StreamManagementView` and `ArkavoView` to align with these changes.

* Rename StreamManagementView to StreamView

Updated the project files to reflect the new name. Modified instances of StreamManagementView in ArkavoView and project.pbxproj accordingly. This change aims to simplify naming and enhance clarity.

* Add stream policies and update view models accordingly

Introduced AdmissionPolicy and InteractionPolicy enums to manage stream access and interaction rules. Updated Stream and related views to incorporate these policies, enhancing the management and creation of streams.