Arkavo v2 UI (#86)
* Add new Discord and Chat views along with ContentView

Introduces `DiscordView` and `ChatView` to enhance messaging and server functionalities. Implements a `ContentView` with a new tabbed navigation. Updated project files to reflect integration and bumped versioning to 2.0.

* Disable unnecessary entitlements for network and downloads.

Removed the "com.apple.security.files.downloads.read-write" and "com.apple.security.network.server" entitlements. This reduces the app's access privileges to improve security and align with the principle of least privilege.

* Update project version to release candidate 0 (rc.0)

Updated the `CURRENT_PROJECT_VERSION` from `beta.4` to `rc.0` in the Xcode project file. This signifies the transition from beta phase to the release candidate stage in preparation for final release.